move console-replay test to its own file
test/cli/inspect/inspect.test.ts has a websocket describe block whose cases
use hardcoded port 6499 and hostname forms that don't resolve everywhere, so
the file carries environment-dependent failures that mask the fail-before /
pass-after signal for this change. Keep the new test alongside the other
per-topic files in test/cli/inspect/.
